feat: added user_service module
